[Feature] #70 Create Entity for Equipment Category
