--- campair/campair/Entity/Dictionary/Content/temp3.swift
@@ -1,8 +1,0 @@
-//
-//  temp3.swift
-//  campair
-//
-//  Created by Mijin CHOI on 2022/06/17.
-//
-
-import Foundation
