+++ campair/campair/Entity/Dictionary/Detail/EquipmentList.swift
@@ -1,0 +1,19 @@
+//
+//  EquipmentSummary.swift
+//  campair
+//
+//  Created by Lee Myeonghwan on 2022/06/17.
+//
+
+import Foundation
+
+struct EquipmentList {
+    let name: String
+    let paintingName: String
+    var paintingURL: String
+    init() {
+        self.name = ""
+        self.paintingName = "임시이미지"
+        self.paintingURL = ""
+    }
+}
