--- campair/campair/Entity/Dictionary/Detail/temp2.swift
@@ -1,8 +1,0 @@
-//
-//  temp2.swift
-//  campair
-//
-//  Created by Mijin CHOI on 2022/06/17.
-//
-
-import Foundation
