--- campair/campair/Entity/Dictionary/EquipmentCollection.swift
@@ -1,12 +1,0 @@
-//
-//  EquipmentCollection.swift
-//  campair
-//
-//  Created by 이가은 on 2022/06/15.
-//
-
-import Foundation
-
-struct EquipmentCollection: Codable {
-    let equipmentGroups: [EquipmentGroup]
-}
