+++ campair/campair/Entity/Dictionary/Main/DictionaryMainCategory.swift
@@ -1,0 +1,19 @@
+//
+//  EquipmentMainCategory.swift
+//  campair
+//
+//  Created by Lee Myeonghwan on 2022/06/17.
+//
+
+import Foundation
+
+struct DictionaryMainCategory {
+    let categoryName: String
+    let paintingName: String
+    var paintingURL: String
+    init() {
+        self.categoryName = ""
+        self.paintingName = "임시이미지"
+        self.paintingURL = ""
+    }
+}
