+++ campair/campair/Entity/Dictionary/Main/DictionaryMainCollection.swift
@@ -1,0 +1,15 @@
+//
+//  temp1.swift
+//  campair
+//
+//  Created by Mijin CHOI on 2022/06/17.
+//
+
+import Foundation
+
+struct DictionaryMainCollection {
+    var dictionaryMainCategory: [DictionaryMainCategory]
+    init() {
+        self.dictionaryMainCategory = [DictionaryMainCategory()]
+    }
+}
