--- campair/campair/Entity/Dictionary/Main/temp1.swift
@@ -1,8 +1,0 @@
-//
-//  temp1.swift
-//  campair
-//
-//  Created by Mijin CHOI on 2022/06/17.
-//
-
-import Foundation
