+++ campair/campair/View/Dictionary/Main/DictionaryMainView.swift
@@ -90,6 +90,7 @@ struct DictionaryMainView_Previews: PreviewProvider {
     }
 }
 
+// TODO: We need to fix it
 struct DictionaryCategoryCollection {
     static var buttons = [
         CategoryButton(id: 0, imageUrl: "paint_1", categoryEnum: Category.tent),
@@ -109,4 +110,56 @@ struct DictionaryCategoryCollection {
         CategoryButton(id: 14, imageUrl: "paint_14", categoryEnum: Category.emergencyKit),
         CategoryButton(id: 15, imageUrl: "paint_15", categoryEnum: Category.toilet)
     ]
+}
+
+enum Category: Int, Codable {
+    case tent
+    case tarp
+    case campingSheet
+    case multiTool
+    case pole
+    case campingTable
+    case campingChair
+    case hammock
+    case sleepingBag
+    case cutlery
+    case mat
+    case campingBox
+    case campingLight
+    case campingTool
+    case emergencyKit
+    case toilet
+    var korean: String {
+        switch self {
+        case .tent: return "텐트"
+        case .tarp: return "타프"
+        case .campingSheet: return "시트"
+        case .multiTool: return "부품"
+        case .pole: return "폴대"
+        case .campingTable: return "테이블"
+        case .campingChair: return "의자"
+        case .hammock: return "해먹"
+        case .sleepingBag: return "침낭"
+        case .cutlery: return "식기"
+        case .mat: return "매트"
+        case .campingBox: return "수납"
+        case .campingLight: return "조명"
+        case .campingTool: return "캠핑공구"
+        case .emergencyKit: return "응급/안전"
+        case .toilet: return "위생"
+        }
+    }
+}
+
+struct EquipmentGroup: Codable {
+    let name: String
+    let equipments: [Equipment]
+}
+
+struct Equipment: Codable {
+    let paintingURL: String
+    let title: String
+    let content: String
+    let pros: [String]
+    let cons: [String]
 }
